Add shared utilities and fix scoping for deployed package selection
- Create shared selectDeployedChange and selectDeployedPackage utilities
- Fix revert command to use scoped deployed packages when --package not supplied
- Add interactive --to flag functionality to verify command
- Both commands now use database-aware package selection with workspace fallback
- Follow established argv passthrough patterns to prevent duplicate prompts
