+++ packages/cli/src/commands/revert.ts
@@ -1,4 +1,4 @@
-import { LaunchQLPackage, LaunchQLMigrate, StatusResult } from '@launchql/core';
+import { LaunchQLPackage } from '@launchql/core';
 import { Logger } from '@launchql/logger';
 import { getEnvOptions } from '@launchql/env';
 import { CLIOptions, Inquirerer, Question } from 'inquirerer';
@@ -6,78 +6,10 @@ import { getPgEnvOptions } from 'pg-env';
 
 import { getTargetDatabase } from '../utils';
 import { selectPackage } from '../utils/module-utils';
+import { selectDeployedChange, selectDeployedPackage } from '../utils/deployed-changes';
 
 const log = new Logger('revert');
 
-async function selectDeployedChange(
-  database: string,
-  argv: Partial<Record<string, any>>,
-  prompter: Inquirerer,
-  log: Logger
-): Promise<string | undefined> {
-  const pgEnv = getPgEnvOptions({ database });
-  const client = new LaunchQLMigrate({
-    host: pgEnv.host,
-    port: pgEnv.port,
-    user: pgEnv.user,
-    password: pgEnv.password,
-    database: pgEnv.database
-  });
-
-  try {
-    let selectedPackage: string;
-    
-    if (argv.package) {
-      selectedPackage = argv.package;
-      log.info(`Using package from CLI: ${selectedPackage}`);
-    } else {
-      const packageStatuses = await client.status();
-      
-      if (packageStatuses.length === 0) {
-        log.warn('No deployed packages found in database');
-        return undefined;
-      }
-
-      const packageAnswer = await prompter.prompt(argv, [{
-        type: 'autocomplete',
-        name: 'package',
-        message: 'Select package to revert from:',
-        options: packageStatuses.map(status => ({
-          name: status.package,
-          value: status.package,
-          description: `${status.totalDeployed} changes, last: ${status.lastChange}`
-        }))
-      }]);
-      selectedPackage = (packageAnswer as any).package;
-    }
-
-    const deployedChanges = await client.getDeployedChanges(database, selectedPackage);
-    
-    if (deployedChanges.length === 0) {
-      log.warn(`No deployed changes found for package ${selectedPackage}`);
-      return undefined;
-    }
-
-    const changeAnswer = await prompter.prompt(argv, [{
-      type: 'autocomplete',
-      name: 'change',
-      message: `Select change to revert to in ${selectedPackage}:`,
-      options: deployedChanges.map(change => ({
-        name: change.change_name,
-        value: change.change_name,
-        description: `Deployed: ${new Date(change.deployed_at).toLocaleString()}`
-      }))
-    }]);
-    const selectedChange = (changeAnswer as any).change;
-
-    return `${selectedPackage}:${selectedChange}`;
-    
-  } catch (error) {
-    log.error('Failed to query deployed changes:', error);
-    log.info('Falling back to non-interactive mode');
-    return undefined;
-  }
-}
 
 const revertUsageText = `
 LaunchQL Revert Command:
@@ -144,7 +76,10 @@ export default async (
 
   let packageName: string | undefined;
   if (recursive && argv.to !== true) {
-    packageName = await selectPackage(argv, prompter, cwd, 'revert', log);
+    packageName = await selectDeployedPackage(database, argv, prompter, log, 'revert');
+    if (!packageName) {
+      packageName = await selectPackage(argv, prompter, cwd, 'revert', log);
+    }
   }
 
   const pkg = new LaunchQLPackage(cwd);
@@ -159,7 +94,7 @@ export default async (
   let target: string | undefined;
   
   if (argv.to === true) {
-    target = await selectDeployedChange(database, argv, prompter, log);
+    target = await selectDeployedChange(database, argv, prompter, log, 'revert');
     if (!target) {
       log.info('No target selected, operation cancelled.');
       return argv;
